Fix detekt LongMethod/TooManyFunctions violations from CI
CI (which does run in this environment, unlike I'd assumed earlier)
caught 7 detekt style violations across the parity-phase commits, all
style/complexity thresholds rather than compile errors:

- DaynestApp.daynestDestinations exceeded LongMethod (63 > 60): extract
  a settingsDestinations() helper, mirroring the existing
  shoppingDestinations() pattern.
- NetworkApiDiModule exceeded TooManyFunctions (12 > 11): split the two
  newest providers (Analytics, Search) into a new GrowthNetworkApiDiModule.
- PrivacyPolicyRoute exceeded LongMethod (65 > 60): extract
  privacyHeader/privacyDataSections/privacyClosingSections helpers.
- SettingsRouteSections.kt exceeded TooManyFunctions (18 > 11): move the
  7 notification/quiet-hours/calendar-feed composables into a new
  SettingsNotificationSections.kt file; loadErrorText made internal so
  both files can share it.
- RecurringGroceriesContent and RecurringGroceryFormDialog both
  exceeded LongMethod (89 and 114 > 60): extract list items, the
  delete-confirmation dialog, form fields, and the list picker into
  their own composables.
- ShoppingListsContent exceeded LongMethod (63 > 60): extract
  active/archived list sections into their own composables.
