+++ android/app/src/main/java/com/daynest/android/core/di/GrowthNetworkApiDiModule.kt
@@ -1,0 +1,22 @@
+package com.daynest.android.core.di
+
+import com.daynest.android.data.analytics.AnalyticsApi
+import com.daynest.android.data.search.SearchApi
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+import retrofit2.Retrofit
+
+@Module
+@InstallIn(SingletonComponent::class)
+object GrowthNetworkApiDiModule {
+    @Provides
+    @Singleton
+    fun provideAnalyticsApi(retrofit: Retrofit): AnalyticsApi = retrofit.create(AnalyticsApi::class.java)
+
+    @Provides
+    @Singleton
+    fun provideSearchApi(retrofit: Retrofit): SearchApi = retrofit.create(SearchApi::class.java)
+}
